backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "qrcode>=8.2",
    "pillow>=11.2.1",
]
